+++ firmware/master/stm32/radio.cpp
@@ -31,7 +31,7 @@ Radio::TxPower Radio::txPower()
     return NRF24L01::instance.txPower();
 }
 
-void Radio::setRetryCount(int hard, int soft)
+void Radio::setRetryCount(uint8_t hard, uint8_t soft)
 {
     NRF24L01::instance.setRetryCount(hard, soft);
 }
